requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
pip==24.0
sqlparse==0.5.3
tzdata==2024.2
